Refactor MountManager for testability, really wish golang had extendable interfaces right about now
