+++ test/sanity/sanity_test.go
@@ -21,6 +21,7 @@ import (
 	compute "google.golang.org/api/compute/v1"
 	gce "sigs.k8s.io/gcp-compute-persistent-disk-csi-driver/pkg/gce-cloud-provider"
 	driver "sigs.k8s.io/gcp-compute-persistent-disk-csi-driver/pkg/gce-pd-csi-driver"
+	mountmanager "sigs.k8s.io/gcp-compute-persistent-disk-csi-driver/pkg/mount-manager"
 )
 
 func TestSanity(t *testing.T) {
@@ -41,10 +42,7 @@ func TestSanity(t *testing.T) {
 		t.Fatalf("Failed to get cloud provider: %v", err)
 	}
 
-	mounter, err := driver.NewFakeSafeFormatAndMount([]error{})
-	if err != nil {
-		t.Fatalf("Failed to get mounter %v", err)
-	}
+	mounter := mountmanager.FakeMounter([]error{})
 
 	//Initialize GCE Driver
 	err = gceDriver.SetupGCEDriver(cloudProvider, mounter, driverName, nodeID, vendorVersion)
